requirements.txt
ultralytics
streamlit
pillow
opencv-python-headless
numpy
pandas
pyyaml
tqdm
matplotlib
